+++ changelog.md
@@ -3,6 +3,9 @@ Liferay Plugin for IntelliJ
 
 All notable changes to this project will be documented in this file.
 
+## 0.0.22
+- Feature: add deprecation check for LPS-104241
+
 ## 0.0.21
 - Feature: add deprecation checks for Liferay 2024.Q1 and Liferay 2024.Q2
 - Feature: add quick fix for non-matching plugin or dependency versions in pom.xml (portal-web, service-builder, rest-builder)